Financial_Sentiment_and_Summarization_API/requirements.txt
pandas>=1.3.0
scikit-learn

newspaper3k
transformers
lxml_html_clean
torch
fastapi
uvicorn
